refactor(help): organize services into embedBuilder/ and actionRow/ subdirectories
